+++ lib/src/controllers/controllers.dart
@@ -21,4 +21,4 @@ part 'navigation/epub_navigation_head_controller.dart';
 part 'navigation/epub_navigation_image_controller.dart';
 part 'navigation/epub_navigation_info_controller.dart';
 part 'navigation/epub_navigation_label_controller.dart';
-part 'navigation/text_image__language_mixin.dart';
+part 'navigation/text_image_language_mixin.dart';
